+++ src/cpp/session/include/session/SessionConstants.hpp
@@ -81,6 +81,8 @@
 
 #define kPackageOutputInPackageFolder     "package-output-to-package-folder"
 
+#define kUseSecureCookiesSessionOption    "session-use-secure-cookies"
+
 // NOTE: literal versions of these are depended upon by the desktop/rsinverse
 // project so they should be updated there as well if they are changed
 #define kLocalUriLocationPrefix           "/rsession-local/"